+++ material/src/components/confirmationDialog/confirmationDialog.component.ts
@@ -3,7 +3,7 @@ import {MatDialogRef} from '@angular/material/dialog';
 import {extend} from '@jscrpt/common';
 
 import {CONFIRMATION_DIALOG_OPTIONS} from '../../misc/tokens';
-import {TITLED_DIALOG_DATA} from '../../misc/interfaces/titledDialog.interface';
+import {TITLED_DIALOG_DATA} from '../../misc/tokens';
 import {ConfirmationDialogOptions} from '../../misc/interfaces/confirmationDialog.interface';
 
 /**