+++ material/src/index.ts
@@ -1,10 +1,14 @@
 export {TitledDialogModule} from './modules/titledDialog.module';
 export {DebugDataCopyClickModule} from './modules/debugDataCopyClick.module';
+export {MovableTitledDialogModule} from './modules/movableTitledDialog.module';
 export {TitledDialogService} from './services/titledDialog/titledDialog.service';
-export {TitledDialogConfig, TITLED_DIALOG_DATA} from './misc/interfaces/titledDialog.interface';
+export {TitledDialogServiceOptions} from './services/titledDialog/titledDialogService.options';
+export {TitledDialogConfig} from './misc/interfaces/titledDialog.interface';
 export {ConfirmationDialogModule} from './modules/confirmationDialog.module';
+export {TitledDialogComponent} from './components/titledDialog/titledDialog.component';
+export {MovableTitledDialogComponent} from './components/movableTitledDialog/movableTitledDialog.component';
 export {ConfirmationDialogComponent} from './components/confirmationDialog/confirmationDialog.component';
 export {ConfirmationDialogDirective} from './directives/confirmationDialog/confirmationDialog.directive';
 export {DebugDataCopyClickDirective} from './directives/debugDataCopyClick/debugDataCopyClick.directive';
 export {ConfirmationDialogOptions, ConfirmationDialogCssClasses} from './misc/interfaces/confirmationDialog.interface';
-export {CONFIRMATION_DIALOG_OPTIONS} from './misc/tokens';
+export {CONFIRMATION_DIALOG_OPTIONS, TITLED_DIALOG_DATA} from './misc/tokens';